+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.13.0] - Unreleased
+
+### Added
+
+- Pretty printer now supports dataclasses
+
+### Fixed
+
+- Fixed Syntax background https://github.com/willmcgugan/rich/issues/1088
+
+### Changed
+
+- Added ws and wss to url highlighter
+
 ## [9.12.4] - 2021-03-01
 
 ### Fixed